add some containers from pcc

+++ kernel/kdebug.h
@@ -1,7 +1,7 @@
 #ifndef KDEBUG_H
 #define KDEBUG_H
 
-#define kassert(tst) kassert_internal((tst), __LINE__, __FILE__, #tst)
+#define kassert(tst) kassert_internal((int)(tst), __LINE__, __FILE__, #tst)
 #define panic(str) panic_internal((str), __LINE__, __FILE__)
 
 void kprintf(const char *fmt, ...) __attribute__ ((format (printf, 1, 2)));
